fixing bug on the funders page
when a ROR link is added, and the manual fields were propagated with the data it was showing as the link was invalid as it wasn't running through the onChange function for the link - creating an error message and not allowing the user to proceed. I've fixed this now by setting the link as valid when a user inputs the data via the ROR method.